Update and rename disable_tamper_protection.py to toggle_tamper_protection.py
Enhanced the functionality to either Enable or Disable Tamper Protection based on Command Line flags. This will make it easier to change Tamper Protection based on preference instead of having to make code changes to the script.

Also fixed some typos and added errors if the '-c' value is not true to false.