Restore subevents dictionary if missing
The key self.current_segment will in a few special cases disappear from
self._events during Process.require_events(). Although there's a change
this only happens if there are no events to be gotten anyway, it's safer to
recreate the missing key and then proceed to get() the required events.