Require query flag
Just running the script without arguments should not lead to automatic resolution of all alerts, this is dangerous.
Discovery of arguments is often use of the script and --help, or -h, and in some cases running the tool without arguments. In this case if the user has a default profile set, all alerts could be set to resolved by mistake.