Add timeout to get_file
The get_file method previously did not support a timeout value.
I have added the timeout values for get_raw_file as arguments to
get_file.

This is a fairly clean way to allow a user of the API to specify a
timeout value and pull large files through cbapi.